Find check remove client

+++ controller/users.js
@@ -120,6 +120,14 @@ exports.removeClient = asyncHandler(async (req, res, next) => {
         id: clientId,
       },
     });
+    const blank = await req.db.blank.findOne({
+      where: {
+        clientId: clientId,
+      },
+    });
+    if (!blank) {
+      throw new MyError("Устгах боломжгүй байна!!");
+    }
     await client.destroy();
   } else if (req.role == "god") {
     const god = await req.db.users.findByPk(req.id);
